+++ netcode/configuration.hh
@@ -19,7 +19,7 @@ struct configuration
   /// @note Possible values are: 4, 8, 16 and 32.
   /// @attention When 4 or 8, length of data handled to the library can be any value. However, for
   /// size 16, length must be a multiple of 2; a multiple of 4 for size 32.
-  std::size_t galois_field_size = 8;
+  std::uint8_t galois_field_size = 8;
 
   /// @brief Tell if the code is systematic or not.
   code code_type = code::systematic;
